plugins/ui/alp-libs/python/pyomopql/requirements.txt
hdbcli==2.12.20
python-dotenv==0.20.0
pandas~=1.3